requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
billiard==4.2.2
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.2
crispy-bootstrap5==0.7
cryptography==46.0.3
cycler==0.12.1
dj-database-url==3.0.1
Django==5.2.8
django-allauth==65.13.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-debug-toolbar==6.1.0
django-extensions==4.1
fonttools==4.60.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.31.0
idna==3.11
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.7
mercadopago==2.3.0
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pandasql==0.7.3
Pillow==10.0.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.4
requests==2.32.5
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
